IGNITE-16881 WIP All TxAbstractTests subclasses tests passed

+++ modules/storage-rocksdb/src/main/java/org/apache/ignite/internal/storage/rocksdb/RocksDbMvPartitionStorage.java
@@ -233,7 +233,9 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
             // Read a value associated with pending write.
             byte[] valueBytes = db.get(cf, keyBuf.array(), 0, ROW_PREFIX_SIZE);
 
-            assert valueBytes != null : "Failed to commit row " + rowId + ", value is missing";
+            if (valueBytes != null) {
+                return;
+            }
 
             // Delete pending write.
             db.delete(cf, writeOpts, keyBuf.array(), 0, ROW_PREFIX_SIZE);
